bot/requirements.txt
python-dotenv==0.21.0
python_telegram_bot==13.15
requests==2.28.1
telegram==0.0.1
